+++ RFL2HMM/Templates/Frontiers/BulletPatternSetParameters.cs
@@ -30,7 +30,7 @@ public class BulletPatternSetParametersClass
                 return Marshal.PtrToStringAnsi((nint)pValue);
             }
 
-            set => pValue = Marshal.StringToHGlobalAnsi(value);
+            set => pValue = (long)Marshal.StringToHGlobalAnsi(value);
         }
 
         public UnmanagedString(string in_value)
