+++ RFL2HMM/Templates/Frontiers/SoundCommonParameter.cs
@@ -55,7 +55,7 @@ public class SoundCommonParameterClass
                 return Marshal.PtrToStringAnsi((nint)pValue);
             }
 
-            set => pValue = Marshal.StringToHGlobalAnsi(value);
+            set => pValue = (long)Marshal.StringToHGlobalAnsi(value);
         }
 
         public UnmanagedString(string in_value)
